improve some functionality: 1. Change field age to Birthday (class Date) 2. Add get by Name (api/employee/name/{name}) 3. Make id of Employee long and mark as @GeneratedValue
change Security access to Employee from Company

+++ src/main/java/io/khasang/archivarius/dao/EmployeeDAO.java
@@ -12,7 +12,7 @@ public interface EmployeeDAO {
 
     void deleteEmployee(Employee employee);
 
-    Employee getEmployeeById(int id);
+    Employee getEmployeeById(long id);
 
     Employee getEmployeeByName(String name);
